requirements.txt
flask==3.0.0
flask-cors==4.0.0
geopandas==0.14.1
pandas==2.1.4
boto3==1.34.0
pyyaml==6.0.1
python-dotenv==1.0.0
Pillow==10.1.0
shapely==2.0.2
psycopg2-binary==2.9.9

